Move Architecture enum to config.py
All our other config enums live there as well so let's colocate
the Architecture enum with them. It also makes the next commit
easier.
