+++ mkosi/distributions/custom.py
@@ -2,12 +2,17 @@
 
 from collections.abc import Sequence
 
+from mkosi.config import Architecture
 from mkosi.distributions import DistributionInstaller
 from mkosi.log import die
 from mkosi.state import MkosiState
 
 
 class Installer(DistributionInstaller):
+    @classmethod
+    def architecture(cls, arch: Architecture) -> str:
+        return str(arch)
+
     @classmethod
     def setup(cls, state: MkosiState) -> None:
         pass
