+++ mkosi/distributions/debian.py
@@ -5,8 +5,8 @@ import tempfile
 from collections.abc import Sequence
 from pathlib import Path
 
-from mkosi.architecture import Architecture
 from mkosi.archive import extract_tar
+from mkosi.config import Architecture
 from mkosi.distributions import Distribution, DistributionInstaller, PackageType
 from mkosi.installer.apt import invoke_apt, setup_apt
 from mkosi.log import die
