+++ mkosi/distributions/openmandriva.py
@@ -3,7 +3,7 @@
 import shutil
 from collections.abc import Sequence
 
-from mkosi.architecture import Architecture
+from mkosi.config import Architecture
 from mkosi.distributions import (
     Distribution,
     DistributionInstaller,
