+++ tests/test_json.py
@@ -8,8 +8,8 @@ from typing import Optional
 
 import pytest
 
-from mkosi.architecture import Architecture
 from mkosi.config import (
+    Architecture,
     BiosBootloader,
     Bootloader,
     Compression,
